refactor: update project and work experience data structure; fix import paths
