+++ src/assets/jsonFiles/workExperience.json
@@ -18,7 +18,7 @@
             "roles": [
                 {
                     "role": "Frontend developer (contract)",
-                    "date": "November 2024 - Present"
+                    "date": "November 2024 - May 2025"
                 }
             ],
             "descriptions": [
